requirements.txt
flask==2.0.1
werkzeug==2.0.3
langchain==0.0.267
faiss-cpu==1.7.4
openai==0.28.0
python-dotenv==0.19.0
requests==2.28.1
tiktoken==0.5.1
tqdm>=4.64.0
torch>=2.0.0
numpy>=1.20.0 